[#1849] reorganized test cases so they will run in CI
