+++ packaging/resource_suite_s3_nocache.py
@@ -1434,45 +1434,3 @@ OUTPUT ruleExecOut
             if os.path.exists(file1):
                 os.unlink(file1)
 
-
-
-class Test_S3_NoCache_V4(Test_S3_NoCache_Base, unittest.TestCase):
-
-    def __init__(self, *args, **kwargs):
-        self.keypairfile='/projects/irods/vsphere-testing/externals/amazon_web_services-CI.keypair'
-        self.s3region='us-east-1'
-        self.s3endPoint='s3.amazonaws.com'
-        self.s3signature_version=4
-        self.s3EnableMPU=1
-        super(Test_S3_NoCache_Base, self).__init__(*args, **kwargs)
-
-class Test_S3_NoCache_EU_Central_1(Test_S3_NoCache_Base, unittest.TestCase):
-    '''
-    This also tests signature V4 with the x-amz-date header.
-    '''
-    def __init__(self, *args, **kwargs):
-        self.keypairfile='/projects/irods/vsphere-testing/externals/amazon_web_services-CI.keypair'
-        self.s3region='eu-central-1'
-        self.s3endPoint='s3.eu-central-1.amazonaws.com'
-        self.s3signature_version=4
-        self.s3EnableMPU=1
-        super(Test_S3_NoCache_Base, self).__init__(*args, **kwargs)
-
-class Test_S3_NoCache_V2(Test_S3_NoCache_Base, unittest.TestCase):
-    def __init__(self, *args, **kwargs):
-        self.keypairfile='/projects/irods/vsphere-testing/externals/amazon_web_services-CI.keypair'
-        self.s3region='us-east-1'
-        self.s3endPoint='s3.amazonaws.com'
-        self.s3signature_version=2
-        self.s3EnableMPU=1
-        super(Test_S3_NoCache_Base, self).__init__(*args, **kwargs)
-
-
-class Test_S3_NoCache_MPU_Disabled(Test_S3_NoCache_Base, unittest.TestCase):
-    def __init__(self, *args, **kwargs):
-        self.keypairfile='/projects/irods/vsphere-testing/externals/amazon_web_services-CI.keypair'
-        self.s3region='us-east-1'
-        self.s3endPoint='s3.amazonaws.com'
-        self.s3signature_version=2
-        self.s3EnableMPU=0
-        super(Test_S3_NoCache_Base, self).__init__(*args, **kwargs)
